pr feedback updates
