+++ needle-slack-connector/.env.example
@@ -1,4 +1,4 @@
-DATABASE_URL="postgresql://postgres:postgres@127.0.0.1:5432/needle-zendesk-connector-dev"
+DATABASE_URL="postgresql://postgres:postgres@127.0.0.1:5432/needle-slack-connector-dev"
 NEEDLE_APP_SECRET=your-app-secret
 NEEDLE_URL=https://needle-ai.com
 NEEDLE_APP_ENV=dev
